+++ compile-tsc
@@ -5,7 +5,9 @@ mkdir -p build
 node_modules/.bin/tsc --noImplicitAny --noEmitOnError --declaration --removeComments --noFallthroughCasesInSwitch --noImplicitReturns --out build/plywood.js \
   src/init.ts \
   \
-  src/dialect/sqlDialect.ts \
+  src/dialect/baseDialect.ts \
+  src/dialect/mySqlDialect.ts \
+  \
   src/datatypes/common.ts \
   src/datatypes/attributeInfo.ts \
   src/datatypes/range.ts \
@@ -16,6 +18,7 @@ node_modules/.bin/tsc --noImplicitAny --noEmitOnError --declaration --removeComm
   src/datatypes/external.ts \
   \
   src/external/druidExternal.ts \
+  src/external/sqlExternal.ts \
   src/external/mySqlExternal.ts \
   \
   src/expressions/baseExpression.ts \
